+++ packages/library-base/src/commonMain/kotlin/io/realm/kotlin/internal/SuspendableWriter.kt
@@ -24,7 +24,6 @@ import io.realm.kotlin.internal.platform.runBlocking
 import io.realm.kotlin.internal.platform.threadId
 import io.realm.kotlin.internal.schema.RealmClassImpl
 import io.realm.kotlin.internal.schema.RealmSchemaImpl
-import io.realm.kotlin.log.RealmLog
 import io.realm.kotlin.query.RealmQuery
 import io.realm.kotlin.types.BaseRealmObject
 import kotlinx.coroutines.CoroutineDispatcher
@@ -78,7 +77,7 @@ internal class SuspendableWriter(private val owner: RealmImpl, val dispatcher: C
     suspend fun updateSchema(schema: RealmSchemaImpl) {
         return withContext(dispatcher) {
             transactionMutex.withLock {
-                RealmLog.debug("Updating schema: $schema")
+                realm.log.debug("Updating schema: $schema")
                 val classPropertyList = schema.classes.map { realmClass: RealmClassImpl ->
                     realmClass.cinteropClass to realmClass.cinteropProperties
                 }