+++ CHANGELOG.md
@@ -6,7 +6,8 @@
 ### Enhancements
 * Multiple processes can now access the same encrypted Realm instead of throwing `Encrypted interprocess sharing is currently unsupported`. (Core Issue [#1845](https://github.com/realm/realm-core/issues/1845))
 * Added a public `RealmLog` class that replaces `AppConfiguration.Builder.log()`. (Issue [#1347](https://github.com/realm/realm-kotlin/pull/1347))
-* Realm logs will now contain more debug information from the underlying database when `LogLevel.TRACE` or below is enabled.
+* Realm logs will now contain more debug information from the underlying database when `LogLevel.DEBUG` or below is enabled.
+* Default log level has been changed from `LogLevel.WARN` to `LogLevel.INFO`.
 * Avoid tracking unreferenced realm versions through the garbage collector. (Issue [#1234](https://github.com/realm/realm-kotlin/issues/1234))
 * [Sync] All tokens, passwords and custom function arguments are now obfuscated by default, even if `LogLevel` is set to DEBUG, TRACE or ALL. (Issue [#410](https://github.com/realm/realm-kotlin/issues/410))
 
@@ -27,7 +28,7 @@
 * Minimum Android SDK: 16.
 
 ### Internal
-* Updated to Realm Core 13.9.0, commit 48e47abd9b25609079785cd7c5316ab4d978ecee.
+* Updated to Realm Core 13.9.1, commit 6c4d43a9d49f8a8a55a1f070865ae1b40cf0c05b.
 
 
 ## 1.7.1 (YYYY-MM-DD)